+++ packages/canon/src/components/Icon/Icon.tsx
@@ -14,21 +14,30 @@
  * limitations under the License.
  */
 
+'use client';
+
 import React from 'react';
 import { useCanon } from '../../contexts/canon';
 import type { IconProps } from './types';
+import clsx from 'clsx';
 
 /** @public */
 export const Icon = (props: IconProps) => {
-  const { name, size = 16 } = props;
+  const { name, size = 16, className, style, ...restProps } = props;
   const { icons } = useCanon();
 
-  const RemixIcon = icons[name] as React.ComponentType<{ className?: string }>;
+  const CanonIcon = icons[name] as React.ComponentType<Omit<IconProps, 'name'>>;
 
-  if (!RemixIcon) {
+  if (!CanonIcon) {
     console.error(`Icon "${name}" not found or is not a valid component.`);
     return <svg />; // Return a default icon or handle the error appropriately
   }
 
-  return <RemixIcon className={`icon-${size}`} />;
+  return (
+    <CanonIcon
+      className={clsx('canon-Icon', className)}
+      style={{ width: size, height: size, ...style }}
+      {...restProps}
+    />
+  );
 };