+++ plugins/catalog-backend-module-github/src/providers/GithubOrgEntityProvider.ts
@@ -14,7 +14,10 @@
  * limitations under the License.
  */
 
-import { SchedulerServiceTaskRunner } from '@backstage/backend-plugin-api';
+import {
+  LoggerService,
+  SchedulerServiceTaskRunner,
+} from '@backstage/backend-plugin-api';
 import { Entity, isGroupEntity } from '@backstage/catalog-model';
 import { Config } from '@backstage/config';
 import {
@@ -47,6 +50,7 @@ import {
 } from '../lib/defaultTransformers';
 import {
   createAddEntitiesOperation,
+  createGraphqlClient,
   createRemoveEntitiesOperation,
   createReplaceEntitiesOperation,
   DeferredEntitiesBuilder,
@@ -56,11 +60,10 @@ import {
   getOrganizationUsers,
   GithubTeam,
 } from '../lib/github';
+import { areGroupEntities, areUserEntities } from '../lib/guards';
 import { assignGroupsToUsers, buildOrgHierarchy } from '../lib/org';
 import { parseGithubOrgUrl } from '../lib/util';
 import { withLocations } from '../lib/withLocations';
-import { areGroupEntities, areUserEntities } from '../lib/guards';
-import { LoggerService } from '@backstage/backend-plugin-api';
 
 const EVENT_TOPICS = [
   'github.membership',
@@ -220,9 +223,11 @@ export class GithubOrgEntityProvider implements EntityProvider {
       await this.credentialsProvider.getCredentials({
         url: this.options.orgUrl,
       });
-    const client = graphql.defaults({
-      baseUrl: this.options.gitHubConfig.apiBaseUrl,
+
+    const client = createGraphqlClient({
       headers,
+      baseUrl: this.options.gitHubConfig.apiBaseUrl!,
+      logger,
     });
 
     const { org } = parseGithubOrgUrl(this.options.orgUrl);