+++ plugins/catalog-unprocessed-entities-common/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-catalog-unprocessed-entities-common
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-permission-common@0.8.4
+
+## 0.0.7-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-permission-common@0.8.4-next.0
+
 ## 0.0.6
 
 ### Patch Changes