+++ plugins/scaffolder-common/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @backstage/plugin-scaffolder-common
 
+## 1.5.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/types@1.2.1
+  - @backstage/catalog-model@1.7.3
+  - @backstage/plugin-permission-common@0.8.4
+
+## 1.5.9-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/types@1.2.1-next.0
+  - @backstage/catalog-model@1.7.3-next.0
+  - @backstage/plugin-permission-common@0.8.4-next.0
+
 ## 1.5.8
 
 ### Patch Changes