+++ plugins/signals/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @backstage/plugin-signals
 
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-plugin-api@1.10.3
+  - @backstage/types@1.2.1
+  - @backstage/core-components@0.16.3
+  - @backstage/theme@0.6.3
+  - @backstage/plugin-signals-react@0.0.9
+
+## 0.0.15-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-plugin-api@1.10.3-next.0
+  - @backstage/types@1.2.1-next.0
+  - @backstage/core-components@0.16.3-next.0
+  - @backstage/plugin-signals-react@0.0.9-next.0
+  - @backstage/theme@0.6.3
+
 ## 0.0.14
 
 ### Patch Changes