+++ plugins/techdocs-addons-test-utils/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @backstage/plugin-techdocs-addons-test-utils
 
+## 1.0.45-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-search-react@1.8.6-next.0
+  - @backstage/plugin-techdocs@1.12.2-next.0
+  - @backstage/core-app-api@1.15.4
+  - @backstage/core-plugin-api@1.10.3
+  - @backstage/integration-react@1.2.3
+  - @backstage/test-utils@1.7.4
+  - @backstage/plugin-catalog@1.26.2-next.0
+  - @backstage/plugin-catalog-react@1.15.2-next.0
+  - @backstage/plugin-techdocs-react@1.2.13
+
+## 1.0.44
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.15.1
+  - @backstage/plugin-techdocs@1.12.1
+  - @backstage/core-plugin-api@1.10.3
+  - @backstage/plugin-catalog@1.26.1
+  - @backstage/core-app-api@1.15.4
+  - @backstage/integration-react@1.2.3
+  - @backstage/test-utils@1.7.4
+  - @backstage/plugin-search-react@1.8.5
+  - @backstage/plugin-techdocs-react@1.2.13
+
+## 1.0.44-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-techdocs@1.12.1-next.1
+  - @backstage/core-plugin-api@1.10.3-next.0
+  - @backstage/plugin-catalog@1.26.1-next.1
+  - @backstage/plugin-catalog-react@1.15.1-next.1
+  - @backstage/plugin-search-react@1.8.5-next.0
+  - @backstage/core-app-api@1.15.4-next.0
+  - @backstage/integration-react@1.2.3-next.0
+  - @backstage/test-utils@1.7.4-next.0
+  - @backstage/plugin-techdocs-react@1.2.13-next.0
+
 ## 1.0.44-next.0
 
 ### Patch Changes